Move classes of the simulated run in their own package
